LayoutTests/imported/w3c/web-platform-tests/html/dom/elements/global-attributes/the-anchor-attribute-004.tentative-expected.txt

Anchor 1
Anchor 2

FAIL the-anchor-attribute-004 assert_equals: Non-HTML elements can use the anchor attribute expected (object) Element node <div id="anchor1">Anchor 1</div> but got (undefined) undefined
